requirements.txt
stanza==1.7.0
neo4j==5.17.0
python-dotenv==1.0.0